+++ invesdwin-nowicket-parent/invesdwin-nowicket-examples/invesdwin-nowicket-examples-mvp-bsgcoach/src/main/java/com/bsgcoach/reports/cor/plannedshippingpercentages/parser/PlannedShippingPercentagesSubReportParserRow.java
@@ -18,112 +18,112 @@ public enum PlannedShippingPercentagesSubReportParserRow {
         @Override
         public void parse(final PlannedShippingPercentages report, final Decimal value) {
             report.setPlannedShippingPercent(CompanyRegion.NorthAmerica, CompanyRegion.NorthAmerica,
-                    new Percent(value, PercentScale.RATE));
+                    new Percent(value.doubleValue(), PercentScale.RATE));
         }
     },
     NaToEa("N.A. Plant to-------------", "E-A") {
         @Override
         public void parse(final PlannedShippingPercentages report, final Decimal value) {
             report.setPlannedShippingPercent(CompanyRegion.NorthAmerica, CompanyRegion.EuropeAfrica,
-                    new Percent(value, PercentScale.RATE));
+                    new Percent(value.doubleValue(), PercentScale.RATE));
         }
     },
     NaToAp("N.A. Plant to-------------", "A-P") {
         @Override
         public void parse(final PlannedShippingPercentages report, final Decimal value) {
             report.setPlannedShippingPercent(CompanyRegion.NorthAmerica, CompanyRegion.AsiaPacific,
-                    new Percent(value, PercentScale.RATE));
+                    new Percent(value.doubleValue(), PercentScale.RATE));
         }
     },
     NaToLa("N.A. Plant to-------------", "L.A.") {
         @Override
         public void parse(final PlannedShippingPercentages report, final Decimal value) {
             report.setPlannedShippingPercent(CompanyRegion.NorthAmerica, CompanyRegion.LatinAmerica,
-                    new Percent(value, PercentScale.RATE));
+                    new Percent(value.doubleValue(), PercentScale.RATE));
         }
     },
     EaToNa("E-A Plant to-------------", "N.A.") {
         @Override
         public void parse(final PlannedShippingPercentages report, final Decimal value) {
             report.setPlannedShippingPercent(CompanyRegion.EuropeAfrica, CompanyRegion.NorthAmerica,
-                    new Percent(value, PercentScale.RATE));
+                    new Percent(value.doubleValue(), PercentScale.RATE));
         }
     },
     EaToEa("E-A Plant to-------------", "E-A") {
         @Override
         public void parse(final PlannedShippingPercentages report, final Decimal value) {
             report.setPlannedShippingPercent(CompanyRegion.EuropeAfrica, CompanyRegion.EuropeAfrica,
-                    new Percent(value, PercentScale.RATE));
+                    new Percent(value.doubleValue(), PercentScale.RATE));
         }
     },
     EaToAp("E-A Plant to-------------", "A-P") {
         @Override
         public void parse(final PlannedShippingPercentages report, final Decimal value) {
             report.setPlannedShippingPercent(CompanyRegion.EuropeAfrica, CompanyRegion.AsiaPacific,
-                    new Percent(value, PercentScale.RATE));
+                    new Percent(value.doubleValue(), PercentScale.RATE));
         }
     },
     EaToLa("E-A Plant to-------------", "L.A.") {
         @Override
         public void parse(final PlannedShippingPercentages report, final Decimal value) {
             report.setPlannedShippingPercent(CompanyRegion.EuropeAfrica, CompanyRegion.LatinAmerica,
-                    new Percent(value, PercentScale.RATE));
+                    new Percent(value.doubleValue(), PercentScale.RATE));
         }
     },
     ApToNa("A-P Plant to-------------", "N.A.") {
         @Override
         public void parse(final PlannedShippingPercentages report, final Decimal value) {
             report.setPlannedShippingPercent(CompanyRegion.AsiaPacific, CompanyRegion.NorthAmerica,
-                    new Percent(value, PercentScale.RATE));
+                    new Percent(value.doubleValue(), PercentScale.RATE));
         }
     },
     ApToEa("A-P Plant to-------------", "E-A") {
         @Override
         public void parse(final PlannedShippingPercentages report, final Decimal value) {
             report.setPlannedShippingPercent(CompanyRegion.AsiaPacific, CompanyRegion.EuropeAfrica,
-                    new Percent(value, PercentScale.RATE));
+                    new Percent(value.doubleValue(), PercentScale.RATE));
         }
     },
     ApToAp("A-P Plant to-------------", "A-P") {
         @Override
         public void parse(final PlannedShippingPercentages report, final Decimal value) {
             report.setPlannedShippingPercent(CompanyRegion.AsiaPacific, CompanyRegion.AsiaPacific,
-                    new Percent(value, PercentScale.RATE));
+                    new Percent(value.doubleValue(), PercentScale.RATE));
         }
     },
     ApToLa("A-P Plant to-------------", "L.A.") {
         @Override
         public void parse(final PlannedShippingPercentages report, final Decimal value) {
             report.setPlannedShippingPercent(CompanyRegion.AsiaPacific, CompanyRegion.LatinAmerica,
-                    new Percent(value, PercentScale.RATE));
+                    new Percent(value.doubleValue(), PercentScale.RATE));
         }
     },
     LaToNa("L.A. Plant-----------------", "N.A.") {
         @Override
         public void parse(final PlannedShippingPercentages report, final Decimal value) {
             report.setPlannedShippingPercent(CompanyRegion.LatinAmerica, CompanyRegion.NorthAmerica,
-                    new Percent(value, PercentScale.RATE));
+                    new Percent(value.doubleValue(), PercentScale.RATE));
         }
     },
     LaToEa("L.A. Plant-----------------", "E-A") {
         @Override
         public void parse(final PlannedShippingPercentages report, final Decimal value) {
             report.setPlannedShippingPercent(CompanyRegion.LatinAmerica, CompanyRegion.EuropeAfrica,
-                    new Percent(value, PercentScale.RATE));
+                    new Percent(value.doubleValue(), PercentScale.RATE));
         }
     },
     LaToAp("L.A. Plant-----------------", "A-P") {
         @Override
         public void parse(final PlannedShippingPercentages report, final Decimal value) {
             report.setPlannedShippingPercent(CompanyRegion.LatinAmerica, CompanyRegion.AsiaPacific,
-                    new Percent(value, PercentScale.RATE));
+                    new Percent(value.doubleValue(), PercentScale.RATE));
         }
     },
     LaToLa("L.A. Plant-----------------", "L.A.") {
         @Override
         public void parse(final PlannedShippingPercentages report, final Decimal value) {
             report.setPlannedShippingPercent(CompanyRegion.LatinAmerica, CompanyRegion.LatinAmerica,
-                    new Percent(value, PercentScale.RATE));
+                    new Percent(value.doubleValue(), PercentScale.RATE));
         }
     };
