+++ invesdwin-nowicket-parent/invesdwin-nowicket-examples/invesdwin-nowicket-examples-mvp-bsgcoach/src/main/java/com/bsgcoach/reports/cor/plantoperations/parser/PlantOperationsSubReportParserRow.java
@@ -69,7 +69,7 @@ public enum PlantOperationsSubReportParserRow {
     PercentOfTotalCapacity("% of Total Capacity", null) {
         @Override
         public void parse(final PlantOperations report, final Decimal value) {
-            report.getPlantCapacity().setPercentOfTotalCapacity(new Percent(value, PercentScale.PERCENT));
+            report.getPlantCapacity().setPercentOfTotalCapacity(new Percent(value.doubleValue(), PercentScale.PERCENT));
         }
     },
     //    Const (for next yr)
@@ -266,7 +266,8 @@ public enum PlantOperationsSubReportParserRow {
     PercentChngFromLastYr("% Chng from Last Yr", null) {
         @Override
         public void parse(final PlantOperations report, final Decimal value) {
-            report.getCompensationDollarsPerWorker().setPercentChngFromLastYr(new Percent(value, PercentScale.PERCENT));
+            report.getCompensationDollarsPerWorker()
+                    .setPercentChngFromLastYr(new Percent(value.doubleValue(), PercentScale.PERCENT));
         }
     },
 
@@ -274,7 +275,8 @@ public enum PlantOperationsSubReportParserRow {
     WageChangePercent("Wage Change (%)", null) {
         @Override
         public void parse(final PlantOperations report, final Decimal value) {
-            report.getProductivityFactors().setWageChangePercent(new Percent(value, PercentScale.PERCENT));
+            report.getProductivityFactors()
+                    .setWageChangePercent(new Percent(value.doubleValue(), PercentScale.PERCENT));
         }
     },
     //    Expected Incent Pay
@@ -295,7 +297,8 @@ public enum PlantOperationsSubReportParserRow {
     TotalCompPercentOfAvg("Total Comp (% of avg)", null) {
         @Override
         public void parse(final PlantOperations report, final Decimal value) {
-            report.getProductivityFactors().setTotalCompPercentOfAvg(new Percent(value, PercentScale.PERCENT));
+            report.getProductivityFactors()
+                    .setTotalCompPercentOfAvg(new Percent(value.doubleValue(), PercentScale.PERCENT));
         }
     },
     //    Best Pract (per wkr)
@@ -324,21 +327,21 @@ public enum PlantOperationsSubReportParserRow {
     PercentChngWage("% Chng--", "Wage") {
         @Override
         public void parse(final PlantOperations report, final Decimal value) {
-            report.getProductivityCalc().setPercentChngWage(new Percent(value, PercentScale.PERCENT));
+            report.getProductivityCalc().setPercentChngWage(new Percent(value.doubleValue(), PercentScale.PERCENT));
         }
     },
     //        BestP
     PercentChngBestP("% Chng--", "BestP") {
         @Override
         public void parse(final PlantOperations report, final Decimal value) {
-            report.getProductivityCalc().setPercentChngBestP(new Percent(value, PercentScale.PERCENT));
+            report.getProductivityCalc().setPercentChngBestP(new Percent(value.doubleValue(), PercentScale.PERCENT));
         }
     },
     //        Opt D
     PercentChngOptD("% Chng--", "Opt D") {
         @Override
         public void parse(final PlantOperations report, final Decimal value) {
-            report.getProductivityCalc().setPercentChngOptD(new Percent(value, PercentScale.PERCENT));
+            report.getProductivityCalc().setPercentChngOptD(new Percent(value.doubleValue(), PercentScale.PERCENT));
         }
     },
     //    New Base (this yr)
@@ -352,14 +355,15 @@ public enum PlantOperationsSubReportParserRow {
     PercentChngIncentive("% Chng--", "Incentive") {
         @Override
         public void parse(final PlantOperations report, final Decimal value) {
-            report.getProductivityCalc().setPercentChngIncentive(new Percent(value, PercentScale.PERCENT));
+            report.getProductivityCalc()
+                    .setPercentChngIncentive(new Percent(value.doubleValue(), PercentScale.PERCENT));
         }
     },
     //        Tot Comp
     PercentChngTotComp("% Chng--", "Tot Comp") {
         @Override
         public void parse(final PlantOperations report, final Decimal value) {
-            report.getProductivityCalc().setPercentChngTotComp(new Percent(value, PercentScale.PERCENT));
+            report.getProductivityCalc().setPercentChngTotComp(new Percent(value.doubleValue(), PercentScale.PERCENT));
         }
     },
     //    Productivity (this yr)
@@ -424,7 +428,8 @@ public enum PlantOperationsSubReportParserRow {
     TqmDollarPerPairCap("TQM----------", "$/pair Cap") {
         @Override
         public void parse(final PlantOperations report, final Decimal value) {
-            report.getCumulatives().getTqm().setDollarPerPairCap(new Percent(value, PercentScale.PERCENT));
+            report.getCumulatives().getTqm().setDollarPerPairCap(
+                    new Percent(value.doubleValue(), PercentScale.PERCENT));
         }
     },
     //    BestP---------  $000s
@@ -445,7 +450,8 @@ public enum PlantOperationsSubReportParserRow {
     BestPDollarPerPairProd("BestP---------", "$/pair Prod") {
         @Override
         public void parse(final PlantOperations report, final Decimal value) {
-            report.getCumulatives().getBestP().setDollarPerPairCap(new Percent(value, PercentScale.PERCENT));
+            report.getCumulatives().getBestP().setDollarPerPairCap(
+                    new Percent(value.doubleValue(), PercentScale.PERCENT));
         }
     },
     //        Employees
@@ -459,7 +465,7 @@ public enum PlantOperationsSubReportParserRow {
     BestPEffectPercent("BestP Effect %", null) {
         @Override
         public void parse(final PlantOperations report, final Decimal value) {
-            report.getMaterialsCosts().setBestPEffectPercent(new Percent(value, PercentScale.PERCENT));
+            report.getMaterialsCosts().setBestPEffectPercent(new Percent(value.doubleValue(), PercentScale.PERCENT));
         }
     },
     //    Per Pair------  Normal
